backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.11.0
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
cachetools==6.2.2
cbor2==5.7.1
certifi==2025.11.12
cffi==2.0.0
channels==4.3.2
charset-normalizer==3.4.4
colorama==0.4.6
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
environ==1.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
hyperlink==21.0.0
idna==3.11
incremental==24.7.2
msgpack==1.1.2
pillow==12.0.0
proto-plus==1.26.1
protobuf==5.29.5
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
PyJWT==2.10.1
pyOpenSSL==25.3.0
pyparsing==3.2.5
requests==2.32.5
rsa==4.9.1
service-identity==24.2.0
setuptools==80.9.0
sniffio==1.3.1
sqlparse==0.5.3
tenacity==9.1.2
tqdm==4.67.1
Twisted==25.5.0
txaio==25.9.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
websockets==15.0.1
zope.interface==8.1.1